requirements.txt
pytorch-lightning==1.6.1
playsound==1.3.0
seaborn==0.11.2
tqdm==4.62.3
torchsummary==1.5.1
halonet-pytorch==0.0.4
matplotlib==3.5.1
gradio==3.1.4
pandas==1.3.4
ruamel.yaml==0.17.21
scikit_learn==1.0.1
scipy==1.8.0
numpy==1.20.0
Pillow==9.0.1
ptflops==0.6.9
timm==0.6.11
SoundFile
PyQt5
librosa
